docs/requirements_for_rtd_aiida.txt
Flask-Cache==0.13.1
Flask-Cors==3.0.7
Flask-HTTPAuth==3.2.4
Flask-RESTful==0.3.7
Flask-SQLAlchemy==2.3.2
Flask==1.0.2
Jinja2==2.10.1
MarkupSafe==1.1.1
PyCifRW==4.2.1; python_version < '3'
PyCifRW==4.4; python_version >= '3'
PyMySQL==0.9.3
PyYAML>=4.2b1
Pygments==2.3.1
SQLAlchemy-Utils==0.33.11
SQLAlchemy==1.3.3
Sphinx==1.8.4
aiida-export-migration-tests==0.7.0
aldjemy==0.9.1
alembic==1.0.7
ase==3.17.0
circus==0.15.0
click-completion==0.5.1
click-config-file==0.5.0
click-spinner==0.1.8
click==7.0
codecov==2.0.15
coverage==4.5.2
django>=1.11.23
docutils==0.14
enum34==1.1.6; python_version<'3.5'
ete3==3.1.1
flask-marshmallow==0.9.0
futures; python_version=='2.7'
graphviz==0.10.1
ipython>=4.0,<6.0
itsdangerous==1.1.0
kiwipy[rmq]==0.5.1
marshmallow-sqlalchemy==0.16.0
mock==2.0.0
monty==2.0.4
numpy==1.16.4,<1.18.0
paramiko==2.6.0
passlib==1.7.1
pg8000<1.13.0
pgtest==1.2.0
pika==1.0.0
plumpy==0.14.2
psutil==5.5.1
pyblake2==1.1.2; python_version<'3.6'
pymatgen<=2018.12.12
pyparsing==2.3.1
pytest-cov==2.6.1
pytest==4.3.0
python-dateutil==2.8.0
python-memcached==1.59
pytz==2018.9
reentry>=1.3.0
seekpath==1.8.4
simplejson==3.16.0
singledispatch>=3.4.0.3; python_version<'3.5'
six==1.12.0
spglib==1.12.2.post0
sphinx-rtd-theme==0.4.3
sphinxcontrib-contentui==0.2.2
sqlalchemy-diff==0.1.3
sqlalchemy-migrate==0.12.0
tabulate==0.8.3
tornado<5.0
typing==3.6.6; python_version<'3.5'
tzlocal==1.5.1
unittest2==1.1.0; python_version<'3.5'
uritools==2.2.0
wrapt==1.11.1
